requirements.txt
aniso8601==7.0.0
asgiref==3.8.1
boto3==1.35.37
botocore==1.35.37
Django==5.1.2
django-filter==24.3
django-graphql-jwt==0.4.0
django-storages==1.14.4
djangorestframework==3.15.2
graphene==2.1.9
graphene-django==2.16.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
jmespath==1.0.1
pillow==11.0.0
promise==2.3
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Rx==1.6.3
s3transfer==0.10.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
urllib3==2.2.3
